fix ups from my review
